telemetry: use SYS_STATUS for RC valid flag
Instead of counting the RC channels we should look at the SYS_STATUS
message. That's because some receivers still send channels even if RC is
lost and they communicate the lost signal using e.g. a low throttle
value.

This change fixes this by switching to SYS_STATUS. Also, we can now get
rid of the timeout around RC_CHANNELS. We keep the subscription to
RC_CHANNELS because we can use it for the signal strength, plus we might
need it later when we add a telemetry field for RC channels.